+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All changes without author are by [Ilya Zverev](https://github.com/Zverik).
 
 * Fixed loading Google scripts after window has finished loading.
 * MapBBCodeProcessor.getLeafletLayers() did not move proprietary layers down. [#43](https://github.com/MapBBCode/mapbbcode/issues/43)
+* `shareTag` option was moved from MapBBCodeUI to MapBBCodeProcessor, allowing for non-standard brackets around `mapid`. [#42](https://github.com/MapBBCode/mapbbcode/issues/42)
 
 ## 1.1.2 (2013-11-29)